requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Authentication & Security
authlib==1.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Task Queue & Scheduling
celery==5.3.4
redis==5.0.1
celery-beat==2.2.1

# HTTP Requests
httpx==0.25.2
aiohttp==3.9.1

# AI/LLM
openai==1.3.7
anthropic==0.7.7

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Environment & Config
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Timezone handling
pytz==2023.3
